+++ src/main/java/com/winterfoodies/winterfoodies_project/dto/product/ProductDto.java
@@ -14,25 +14,25 @@ import lombok.ToString;
 public class ProductDto {
 
     @ApiModelProperty(example = "상품 id" )
-    private Long productId;
+    private double productId;
 
     @ApiModelProperty(example = "상품명" )
     private String productName;
 
     @ApiModelProperty(example = "상품 수량" )
-    private Long quantity;
+    private double quantity;
 
     @ApiModelProperty(example = "메시지" )
     private String message;
 
     @ApiModelProperty(example = "가게명" )
-    private Long storeId;
+    private double storeId;
 
     @ApiModelProperty(example = "500", value = "각 상품당 가격")
-    private Long pricePerProduct;
+    private double pricePerProduct;
 
     @ApiModelProperty(example = "3000", value = "각 메뉴별 총 주문금액")
-    private Long subTotalAmount;
+    private double subTotalAmount;
 
     public ProductDto() {
     }
