+++ src/main/java/com/winterfoodies/winterfoodies_project/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    Optional<Product> findById(Long productId);
+    Optional<Product> findById(double productId);
 
 
 }
